tech(TTB-060): Moved gameUtils.js to new utils directory
